fix: 마이페이지 FSD 적용
