+++ src/app/(afterLogin)/mypage/page.tsx
@@ -1,199 +1,15 @@
-'use client';
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { useLogout } from '@/app/shared/hooks/queries/useAuth';
-import { useAuthStore } from '@/app/shared/store/useAuthStore';
-import LogoutConfirm from '@/app/features/mypage/components/LogoutConfirm';
-import Button from '@/app/shared/ui/Button';
+import MyPageContainer from '@/app/features/mypage/components/MyPageContainer';
+import { Metadata } from 'next';
 
-const menuItems = [
-  {
-    id: 'my-reviews',
-    title: '내 리뷰',
-    icon: '📝',
-    path: '/mypage/reviews',
-  },
-  {
-    id: 'baropot',
-    title: '바로팟',
-    icon: '⚡',
-    path: '/mypage/baropot',
-  },
-  {
-    id: 'bookmarks',
-    title: '찜한곳',
-    icon: '💙',
-    path: '/mypage/bookmarks',
-  },
-  {
-    id: 'visited',
-    title: '방문기록(지도)',
-    icon: '📍',
-    path: '/mypage/visited',
-  },
-  {
-    id: 'settings',
-    title: '설정',
-    icon: '⚙️',
-    path: '/mypage/settings',
-  },
-  {
-    id: 'history',
-    title: '히스토리',
-    icon: '📊',
-    path: '/mypage/history',
-  },
-];
+export const metadata: Metadata = {
+  title: '바로고 | 마이페이지',
+  description: '마이페이지에서 내 정보를 관리해보세요.',
+};
 
 export default function MyPage() {
-  const router = useRouter();
-  const { user } = useAuthStore();
-  const logout = useLogout();
-  const [showLogoutModal, setShowLogoutModal] = useState(false);
-
-  const [userStats] = useState({
-    reviews: 12,
-    baropots: 12,
-    bookmarks: 12,
-  });
-
-  const handleMenuClick = (path: string) => {
-    router.push(path);
-  };
-
   return (
     <div className="min-h-screen bg-[#E6EEF5] pb-24">
-      {/* 헤더 */}
-      <div className="sticky top-0 z-40 border-b border-gray-200 bg-white">
-        <div className="flex items-center px-4 py-3">
-          <button
-            onClick={() => router.back()}
-            className="rounded-lg p-2 text-gray-600 hover:bg-gray-100"
-          >
-            ←
-          </button>
-          <h1 className="flex-1 text-center text-lg font-semibold text-[#2B2B2B]">
-            마이페이지
-          </h1>
-          <div className="w-10"></div>
-        </div>
-      </div>
-
-      <div className="space-y-6 px-4 py-6">
-        {/* 프로필 섹션 */}
-        <div className="rounded-xl bg-white p-6 shadow-sm">
-          <div className="flex items-center space-x-4">
-            {/* 프로필 이미지 - 실제 사용자 이름 사용 */}
-            <div className="flex h-16 w-16 items-center justify-center rounded-full bg-[#1C4E80]">
-              <span className="text-2xl font-bold text-white">
-                {user?.name?.charAt(0) || 'U'}
-              </span>
-            </div>
-
-            {/* 사용자 정보 - 실제 데이터 사용 */}
-            <div className="flex-1">
-              <h2 className="mb-1 text-xl font-semibold text-[#2B2B2B]">
-                {user?.name || '사용자'}
-              </h2>
-              <p className="text-sm text-[#8A8A8A]">
-                {user?.email || '맛집 탐험가'}
-              </p>
-            </div>
-          </div>
-
-          {/* 통계 */}
-          <div className="mt-6 grid grid-cols-3 gap-4">
-            <div className="text-center">
-              <div className="mb-1 text-2xl font-bold text-[#1C4E80]">
-                {userStats.reviews}
-              </div>
-              <div className="text-sm text-[#8A8A8A]">리뷰</div>
-            </div>
-            <div className="text-center">
-              <div className="mb-1 text-2xl font-bold text-[#1C4E80]">
-                {userStats.baropots}
-              </div>
-              <div className="text-sm text-[#8A8A8A]">바로팟</div>
-            </div>
-            <div className="text-center">
-              <div className="mb-1 text-2xl font-bold text-[#1C4E80]">
-                {userStats.bookmarks}
-              </div>
-              <div className="text-sm text-[#8A8A8A]">찜한곳</div>
-            </div>
-          </div>
-        </div>
-
-        {/* 메뉴 그리드 */}
-        <div className="grid grid-cols-2 gap-4">
-          {menuItems.map((item) => (
-            <button
-              key={item.id}
-              onClick={() => handleMenuClick(item.path)}
-              className="rounded-xl bg-white p-6 text-center shadow-sm transition-all duration-200 hover:shadow-md"
-            >
-              <div className="mb-3 text-3xl">{item.icon}</div>
-              <div className="font-medium text-[#2B2B2B]">{item.title}</div>
-            </button>
-          ))}
-        </div>
-
-        {/* 추가 정보 섹션 */}
-        <div className="rounded-xl bg-white p-4 shadow-sm">
-          <h3 className="mb-3 font-semibold text-[#2B2B2B]">최근 활동</h3>
-          <div className="space-y-3">
-            <div className="flex items-center space-x-3">
-              <div className="flex h-8 w-8 items-center justify-center rounded-full bg-blue-100">
-                <span className="text-sm text-blue-600">⚡</span>
-              </div>
-              <div className="flex-1">
-                <p className="text-sm text-[#2B2B2B]">
-                  홍대 맛집 바로팟에 참여했어요
-                </p>
-                <p className="text-xs text-[#8A8A8A]">2시간 전</p>
-              </div>
-            </div>
-
-            <div className="flex items-center space-x-3">
-              <div className="flex h-8 w-8 items-center justify-center rounded-full bg-orange-100">
-                <span className="text-sm text-orange-600">📝</span>
-              </div>
-              <div className="flex-1">
-                <p className="text-sm text-[#2B2B2B]">
-                  석압생소금구이 리뷰를 작성했어요
-                </p>
-                <p className="text-xs text-[#8A8A8A]">1일 전</p>
-              </div>
-            </div>
-
-            <div className="flex items-center space-x-3">
-              <div className="flex h-8 w-8 items-center justify-center rounded-full bg-red-100">
-                <span className="text-sm text-red-600">💙</span>
-              </div>
-              <div className="flex-1">
-                <p className="text-sm text-[#2B2B2B]">홀리스 타코를 찜했어요</p>
-                <p className="text-xs text-[#8A8A8A]">3일 전</p>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* 개선된 로그아웃 버튼 */}
-        <Button
-          text={logout.isPending ? '로그아웃 중...' : '로그아웃'}
-          onClick={() => setShowLogoutModal(true)}
-          disabled={logout.isPending}
-          className="w-full rounded-xl py-4 font-medium"
-        />
-      </div>
-
-      {showLogoutModal && (
-        <LogoutConfirm
-          isOpen={showLogoutModal}
-          onClose={() => setShowLogoutModal(false)}
-          logout={logout}
-        />
-      )}
+      <MyPageContainer />
     </div>
   );
 }
